requirements.txt
discord.py>=2.6.2
python-dotenv>=1.0.1
requests>=2.32.5